+++ CHANGES.md
@@ -1,5 +1,7 @@
 # 0.2
 - Drop support for Python 2
+- Show available commands
+- Show script errors
 
 # 0.1
 - Prepare for public release